Feature/improved options

+++ IO/SpriteSheetIO.cpp
@@ -1,5 +1,8 @@
 #include <iostream>
 #include "SpriteSheetIO.h"
+#include "../logging/LoggerTags.hpp"
+
+namespace logger = LoggerTags;
 
 /**
  * Validates the given path and sets the inFilePath member variable. Note, the member variable is set even if invalid.
@@ -16,13 +19,13 @@ bool SpriteSheetIO::setInPath(const std::string& pathName, bool shouldBePNG, boo
     inFilePath_ = fs::path(pathName).make_preferred();
 
     if (! fs::exists(inFilePath_)) {
-        std::cout << "[ERROR] The provided input directory does not exist:\n\t\t" << inFilePath_ << "\n";
+        std::cout << logger::error << "The provided input directory does not exist:\n\t\t" << inFilePath_ << "\n";
         return false;
     }
 
     // not a directory and not a png? that's an error
     if (! (shouldBePNG || fs::is_directory(inFilePath_))) {
-        std::cout << "[ERROR] The provided input directory is not a png file or folder:\n\t\t" << inFilePath_ << "\n";
+        std::cout << logger::error << "The provided input directory is not a png file or folder:\n\t\t" << inFilePath_ << "\n";
         return false;
     }
 
@@ -47,21 +50,29 @@ bool SpriteSheetIO::setInPath(const std::string& pathName, bool shouldBePNG, boo
 bool SpriteSheetIO::setOutPath(const std::string &pathName) {
     outFilePath_ = fs::path(pathName).make_preferred();
 
+    if(! fs::exists(outFilePath_)) {
+        std::cout << logger::warn << "The provided output directory does not exist.\n\t\t" << outFilePath_ << "\n";
+        std::cout << logger::warn << "This directory will be created.\n";
+        std::error_code ec;
+        fs::create_directory(outFilePath_, ec);
+
+        if (ec.value() != 0) return false;
+    }
+
     if (! fs::is_directory(outFilePath_)) {
         outFilePath_ = outFilePath_.parent_path();
     }
 
-    if(! fs::exists(outFilePath_)) {
-        std::cout << "[ERROR] The provided output directory does not exist.\n\t\t" << outFilePath_ << "\n";
-        return false;
-    }
-
     if (! fs::is_directory(outFilePath_)) {
-        std::cout << "[ERROR] The output directory must be a folder.\n\t\t" << outFilePath_ << "\n";
+        std::cout << logger::error << "The output directory must be a folder.\n\t\t" << outFilePath_ << "\n";
         return false;
     }
 
     return true;
+}
+
+void SpriteSheetIO::setIOOptions(const IOOptions &opts) {
+    IOOpts_ = opts;
 }
 
 /**
@@ -113,20 +124,20 @@ void SpriteSheetIO::saveSplits(SpriteSplittingData& ssd, std::basic_ostream<char
     std::error_code ec;
     bool cleanedFolder = createCleanDirectory(folderName, ec);
     if (! cleanedFolder || ec.value() != 0) {
-        outStream << "[ERROR] Failed to create folder " << folderName << "\n\t\t" << ec << "\n";
+        outStream << logger::threaded_error << "Failed to create folder " << folderName << "\n\t\t" << ec << "\n";
         ssd.stats.n_save_error += ssd.spriteCount; // mark every sprite as failed.
         return;
     }
 
     switch (ssd.sheetType) {
         case SpriteSheetType::OBJECT:
-            saveObjectSplits(ssd, folderName);
+            saveObjectSplits(ssd, folderName, outStream);
             break;
         case SpriteSheetType::CHARACTER:
-            saveCharSplits(ssd, folderName);
+            saveCharSplits(ssd, folderName, outStream);
             break;
         default: // did you add a new SpriteSheetType?
-            outStream << "[ERROR]: Unknown SpriteSheetType " << ssd.sheetType << "\n";
+            outStream << logger::threaded_error << "Unknown SpriteSheetType " << ssd.sheetType << "\n";
             exit(-1);
     }
 }
@@ -147,7 +158,7 @@ void SpriteSheetIO::saveSplits(SpriteSplittingData& ssd, std::basic_ostream<char
  *            originalFileName: name of the SpriteSheet the splits originate from
  *            stats: stat tracking object
  */
-void SpriteSheetIO::saveObjectSplits(SpriteSplittingData& ssd, const std::string& folderName) const {
+void SpriteSheetIO::saveObjectSplits(SpriteSplittingData &ssd, const std::string &folderName, std::basic_ostream<char>& outStream) const {
 
     auto* sprite = new unsigned char[ssd.spriteSize * ssd.spriteSize * 4];
     int skippedSprites = 0;
@@ -168,8 +179,10 @@ void SpriteSheetIO::saveObjectSplits(SpriteSplittingData& ssd, const std::string
         if (transparent) {
             skippedSprites++;
         } else {
+            // subtract from the index the amount of alpha sprites we ignored, if this indexing method is user specified.
+            int index = i - (IOOpts_.subtractAlphaFromIndex ? skippedSprites : 0);
             // unsigned char* sprite is now holding a spriteSize * spriteSize * 4 byte sprite. Finally!
-            bool error = saveObjectSprite(sprite, i - skippedSprites, ssd.spriteSize, ssd.lodeState, folderName);
+            bool error = saveObjectSprite(sprite, index, ssd.spriteSize, ssd.lodeState, folderName, outStream);
             ssd.stats.n_save_error +=   error;
             ssd.stats.n_success +=      ! error;
         }
@@ -191,17 +204,24 @@ void SpriteSheetIO::saveObjectSplits(SpriteSplittingData& ssd, const std::string
  *
  * @return whether an error ocurred.
  */
-bool SpriteSheetIO::saveObjectSprite(const unsigned char* sprite, int index, unsigned int spriteSize, lodepng::State& lodeState, const std::string& folderName) const {
-    bool error;
+bool SpriteSheetIO::saveObjectSprite(const unsigned char* sprite, int index, unsigned int spriteSize, lodepng::State& lodeState, const std::string& folderName, std::basic_ostream<char>& outStream) const {
+    unsigned int error;
     std::string fileName = std::to_string(index) + ".png";
     std::vector<unsigned char> encodedPixels;
 
     error = lodepng::encode(encodedPixels, sprite, spriteSize, spriteSize, lodeState);
+    checkLodePNGErrorCode(error, outStream);
     if (!error) {
-        error = lodepng::save_file(encodedPixels, (outFilePath_/folderName/fileName).string());
+        std::filesystem::path outPath = outFilePath_;
+        if (IOOpts_.useSubFolders) { // insert a subfolder in the directory if specified by options.
+            outPath /= folderName;
+        }
+        outPath /= fileName;
+        error = lodepng::save_file(encodedPixels, outPath.string());
+        checkLodePNGErrorCode(error, outStream);
     }
 
-    return error;
+    return static_cast<bool>(error);
 }
 
 /**
@@ -218,7 +238,7 @@ bool SpriteSheetIO::saveObjectSprite(const unsigned char* sprite, int index, uns
  *            stats: stat tracking object
  *
  */
-void SpriteSheetIO::saveCharSplits(SpriteSplittingData& ssd, const std::string& folderName) const {
+void SpriteSheetIO::saveCharSplits(SpriteSplittingData& ssd, const std::string& folderName, std::basic_ostream<char>& outStream) const {
     // for char sheets, one row = one character. If there exists invisible frames on that row (but not all are invisible),
     // then that is perfectly valid. For example, pet skins without attack frames.
     // I suspect Exalt still expects full alpha frames to slot into e.g. a pets attack frames.
@@ -247,8 +267,10 @@ void SpriteSheetIO::saveCharSplits(SpriteSplittingData& ssd, const std::string& 
             if (charSpritesAreAlpha(charSprites, ssd.spriteSize, sprite_4)) {
                 skippedSprites++;
             } else {
+                // subtract from the index the amount of alpha sprites we ignored, if this indexing method is user specified.
+                int index = (i / SPRITES_PER_CHAR) - (IOOpts_.subtractAlphaFromIndex ? skippedSprites : 0);
                 // unsigned char** charSprites is now holding a chars' sprites. Finally!
-                unsigned int errors = saveCharSprites(charSprites, (i / SPRITES_PER_CHAR) - skippedSprites, ssd.spriteSize, ssd.lodeState, folderName);
+                unsigned int errors = saveCharSprites(charSprites, index, ssd.spriteSize, ssd.lodeState, folderName, outStream);
                 ssd.stats.n_save_error += errors;
                 ssd.stats.n_success += static_cast<unsigned int>(SPRITES_PER_CHAR) - errors;
             }
@@ -275,8 +297,8 @@ void SpriteSheetIO::saveCharSplits(SpriteSplittingData& ssd, const std::string& 
  *
  * @return number of errors that occurred.
  */
-unsigned int SpriteSheetIO::saveCharSprites(unsigned char *sprites [SPRITES_PER_CHAR], int index, unsigned int spriteSize, lodepng::State &lodeState, const std::string &folderName) const {
-    bool error; // handily casts away error codes from lodePNG to just "error Y/N". this is intended.
+unsigned int SpriteSheetIO::saveCharSprites(unsigned char *sprites [SPRITES_PER_CHAR], int index, unsigned int spriteSize, lodepng::State &lodeState, const std::string &folderName, std::basic_ostream<char>& outStream) const {
+    unsigned int error;
     unsigned int errorCount = 0;
     std::string baseFileName = std::to_string(index) + '_';
 
@@ -289,16 +311,22 @@ unsigned int SpriteSheetIO::saveCharSprites(unsigned char *sprites [SPRITES_PER_
 
         std::vector<unsigned char> encodedPixels;
         error = lodepng::encode(encodedPixels, sprites[spriteIndex], width, spriteSize, lodeState);
+        checkLodePNGErrorCode(error, outStream);
 
         if (!error) {
-            error = lodepng::save_file(encodedPixels, (outFilePath_ / folderName / fileName).string());
+            std::filesystem::path outPath = outFilePath_;
+            if (IOOpts_.useSubFolders) { // insert a subfolder in the directory if specified by options.
+                outPath /= folderName;
+            }
+            outPath /= fileName;
+            error = lodepng::save_file(encodedPixels, outPath.string());
+            checkLodePNGErrorCode(error, outStream);
         }
 
-        errorCount += error;
+        errorCount += error == 0;
     }
 
     return errorCount;
-
 }
 
 /**
@@ -336,7 +364,6 @@ std::string SpriteSheetIO::folderNameFromSheetName(const std::string& sheetPath,
     std::string fileName = std::move(fs::path(sheetPath).filename().string());
     // lowercase conversion lambda
     auto toLower = [](const std::string& s)->std::string { std::string o; for (const auto& c : s) o.push_back(static_cast<char>(std::tolower(c))); return o; };
-    size_t index;
     std::string specifier;
     switch(type) {
         case SpriteSheetType::OBJECT: {
@@ -351,37 +378,82 @@ std::string SpriteSheetIO::folderNameFromSheetName(const std::string& sheetPath,
     }
 
     std::string lowerName {std::move(toLower(fileName))};
-    index = lowerName.find(specifier);
 
-    if (index == 0) { // some 2010 sheets start with their specifier instead of at the end. Treat those as if no specifier was found. Otherwise, multiple threads would write to e.g. a folder 'chars_16'. This usually results in an error.
-        index = (size_t) -1;
-    } else if (index != (size_t) -1) { // now have the index to the start of specifier, but want to include dimension number as well.
-        index = index + specifier.size(); // (only do this now in case of not found overflow) index points to end of specifier
-        index = lowerName.find('x', index+1); // +1: includes self, what if a future specifier ends in 'x'? Dimensions are at least one char so this is OK.
+    size_t sheet_size_specifier_start_index = lowerName.find(specifier);
+    size_t sheet_name_end_index = std::string::npos;
+
+    switch (sheet_size_specifier_start_index) {
+        case 0:
+        case std::string::npos:
+            break; // cannot deduce a folder name with name, type and dimension specifier.
+        default:
+            // Set the dimension specifier start index to the first character past the size specifier.
+            size_t dimension_specifier_start_index = sheet_size_specifier_start_index + specifier.size();
+            // Set the dimension specifier end index, start searching 1 char past the specifier. Dimensions are at least one char so this is OK.
+            size_t dimension_specifier_end_index = lowerName.find('x', dimension_specifier_start_index + 1);
+
+            // If there is indeed an 'x', assume this is a size specifier.
+            if (dimension_specifier_end_index != std::string::npos) {
+                // This value typically is 1 or 2. (e.g. values 8x8 or 16x16, where start_index points to the first char, and end_index to the 'x' char)
+                size_t dimension_specifier_n_char = dimension_specifier_end_index - dimension_specifier_start_index;
+                // (0) starting from 'start of dimension specifier':
+                // (1) go forward the dimension char amount
+                // (2) go forward 1, to get past the 'x'
+                // (3) go forward the dimension char amount.
+                // This results in an index pointing to the character past the dimension specifier.
+                // The expression in these steps is rewritten to 1 + (2 x size)
+
+                // Example: RandomNameObjects16x16
+                //                           ^ ^^ ^
+                //                           0 12 3
+                sheet_name_end_index = dimension_specifier_start_index + 1 + (2 * dimension_specifier_n_char);
+
+                if (sheet_name_end_index > fileName.size() - 4) {
+                    // this end index points past the fileName (minus .png) !
+                    // This could happen if there was no dimension specifier after all, e.g.
+                    // RandomNameObjects16x      (no number after x)
+                    sheet_name_end_index = std::string::npos;
+                }
+            }
     }
 
-    if (index == (size_t) -1) { // no type specifier? at least split off the .png and suggest that as folder name.
+    if (sheet_name_end_index == std::string::npos) {
+        // no valid type specifier? at least split off the .png and suggest that as folder name.
         return fileName.substr(0, fileName.size() - 4);
     } else {
-        return fileName.substr(0, index);
+        return fileName.substr(0, sheet_name_end_index);
     }
 }
 
 /**
  * Cleans out or creates the directory given by appending the given dir to the outFilePath_
+ *
+ * Depending on the useSubFolders option, either the subdirectory is cleaned, or nothing is done at all:
+ * If a shared folder for multiple sheets gets cleaned, one file's splits would be saved,
+ * only for a future split to empty this folder again and put its splits in.
+ *
  * @param dir the folder(s) to append to outFilePath
  * @return whether the operation was successful
  */
 bool SpriteSheetIO::createCleanDirectory(const std::string& dir, std::error_code& ec) const noexcept {
-    if (fs::exists(outFilePath_/dir)) {
-        fs::remove_all(outFilePath_/dir, ec);
-    }
+     if (IOOpts_.useSubFolders) {
+        if (fs::exists(outFilePath_ / dir)) {
+            fs::remove_all(outFilePath_ / dir, ec);
+        }
 
-    if (ec.value() != 0) return false;
+        if (ec.value() != 0) return false;
 
-    return fs::create_directory(outFilePath_/dir, ec);
+        return fs::create_directory(outFilePath_ / dir, ec);
+     } else { // shared folder case, do not delete other threads' work!!
+         return true;
+     }
 }
 
 SpriteSheetIO::~SpriteSheetIO() {
     delete directoryIterator_;
+}
+
+// Print an error, if and only if the lodePNG error code returned is non-zero.
+void SpriteSheetIO::checkLodePNGErrorCode(unsigned int lode_code, std::basic_ostream<char>& outStream) {
+    if (lode_code) outStream << logger::error << "LodePNG error: " << lodepng_error_text(lode_code) << ".\n";
 }
